third_party/blink/web_tests/platform/generic/editing/deleting/forward-delete-at-table-cell-ending-expected.txt

This test verifies that forward delete at the end of a table-cell is a no-op. The first dump verifies that the caret stays after 'abc'. The second dump verifies that the caret stays after 'def', and doesn't select the nested table.

Dump of markup 1:
| <table>
|   border="1"
|   <tbody>
|     <tr>
|       <td>
|         id="first"
|         "abc<#selection-caret>"
|       <td>
|         id="second"
|         "def"
|       <td>
|         <table>
|           <tbody>
|             <tr>
|               <td>
|                 "ghi"
|               <td>
|                 "jkl"

Dump of markup 2:
| <table>
|   border="1"
|   <tbody>
|     <tr>
|       <td>
|         id="first"
|         "abc"
|       <td>
|         id="second"
|         "def<#selection-caret>"
|       <td>
|         <table>
|           <tbody>
|             <tr>
|               <td>
|                 "ghi"
|               <td>
|                 "jkl"
